Cast switchpoints to int BEFORE getting the mode

+++ additional_analyses/blech_identity_palatability_switch.py
@@ -218,8 +218,8 @@ for i in range(num_trials):
 		# 2.) Switchpoints (averaged over the last 100k samples, skipping 100 samples at a time)
 		#start = int(np.mean(tr[-100000::100]['t1'][:, i]))
 		#end = int(np.mean(tr[-100000::100]['t2'][:, i]))
-		start = int(Counter(tr[-100000::100]['t1'][:, i]).most_common()[0][0])
-		end = int(Counter(tr[-100000::100]['t2'][:, i]).most_common()[0][0])
+		start = int(Counter(tr[-100000::100]['t1'][:, i].astype('int')).most_common()[0][0])
+		end = int(Counter(tr[-100000::100]['t2'][:, i].astype('int')).most_common()[0][0])
 		this_switchpoints.append([start, end])
 		# 3.) Palatability rank
 		this_pal.append(palatability[0, i])
@@ -281,8 +281,8 @@ for i in range(num_trials):
 		# 2.) Switchpoints (averaged over the last 100k samples, skipping 100 samples at a time)
 #		start = int(np.mean(tr[-100000::100]['t1'][:, i]))
 #		end = int(np.mean(tr[-100000::100]['t2'][:, i]))
-		start = int(Counter(tr[-100000::100]['t1'][:, i]).most_common()[0][0])
-		end = int(Counter(tr[-100000::100]['t2'][:, i]).most_common()[0][0])
+		start = int(Counter(tr[-100000::100]['t1'][:, i].astype('int')).most_common()[0][0])
+		end = int(Counter(tr[-100000::100]['t2'][:, i].astype('int')).most_common()[0][0])
 		this_switchpoints.append([start, end])
 		# 3.) Palatability rank
 		this_pal.append(palatability[1, i])
@@ -344,8 +344,8 @@ for i in range(num_trials):
 		# 2.) Switchpoints (averaged over the last 100k samples, skipping 100 samples at a time)
 #		start = int(np.mean(tr[-100000::100]['t1'][:, i]))
 #		end = int(np.mean(tr[-100000::100]['t2'][:, i]))
-		start = int(Counter(tr[-100000::100]['t1'][:, i]).most_common()[0][0])
-		end = int(Counter(tr[-100000::100]['t2'][:, i]).most_common()[0][0])
+		start = int(Counter(tr[-100000::100]['t1'][:, i].astype('int')).most_common()[0][0])
+		end = int(Counter(tr[-100000::100]['t2'][:, i].astype('int')).most_common()[0][0])
 		this_switchpoints.append([start, end])
 		# 3.) Palatability rank
 		this_pal.append(palatability[2, i])
@@ -407,8 +407,8 @@ for i in range(num_trials):
 		# 2.) Switchpoints (averaged over the last 100k samples, skipping 100 samples at a time)
 #		start = int(np.mean(tr[-100000::100]['t1'][:, i]))
 #		end = int(np.mean(tr[-100000::100]['t2'][:, i]))
-		start = int(Counter(tr[-100000::100]['t1'][:, i]).most_common()[0][0])
-		end = int(Counter(tr[-100000::100]['t2'][:, i]).most_common()[0][0])
+		start = int(Counter(tr[-100000::100]['t1'][:, i].astype('int')).most_common()[0][0])
+		end = int(Counter(tr[-100000::100]['t2'][:, i].astype('int')).most_common()[0][0])
 		this_switchpoints.append([start, end])
 		# 3.) Palatability rank
 		this_pal.append(palatability[3, i])
